requirements.txt
GitPython==3.1.40
requests==2.31.0
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2
